+++ src/scripts/src/actions/groups.js
@@ -1,0 +1,8 @@
+/* eslint-disable import/prefer-default-export */
+export const addGroup = (value, id, active, disabled) => ({
+  type: 'ADD_GROUP',
+  value,
+  id,
+  active,
+  disabled,
+});
