+++ src/scripts/src/actions/misc.js
@@ -1,0 +1,4 @@
+/* eslint-disable import/prefer-default-export */
+export const clearAll = () => ({
+  type: 'CLEAR_ALL',
+});
