+++ src/scripts/src/actions/misc.test.js
@@ -1,0 +1,14 @@
+import { expect } from 'chai';
+import * as actions from './misc';
+
+describe('misc actions', () => {
+  describe('clearAll action', () => {
+    it('returns CLEAR_ALL action', () => {
+      const expectedAction = {
+        type: 'CLEAR_ALL',
+      };
+
+      expect(actions.clearAll()).to.eql(expectedAction);
+    });
+  });
+});
